Moved all the visitors into a single package.
