+++ src/main/java/dte/employme/reward/Reward.java
@@ -2,7 +2,7 @@ package dte.employme.reward;
 
 import org.bukkit.entity.Player;
 
-import dte.employme.reward.visitor.RewardVisitor;
+import dte.employme.visitors.reward.RewardVisitor;
 
 public interface Reward
 {
