+++ src/main/java/dte/employme/visitors/goal/GoalInventoryDescriptor.java
@@ -1,16 +1,15 @@
-package dte.employme.board.inventory;
+package dte.employme.visitors.goal;
 
 import static dte.employme.utils.ChatColorUtils.colorize;
 
 import org.bukkit.inventory.ItemStack;
 
 import dte.employme.goal.ItemGoal;
-import dte.employme.goal.visitor.GoalVisitor;
 import dte.employme.utils.EnumUtils;
 
-public class GoalDescriptor implements GoalVisitor<String>
+public class GoalInventoryDescriptor implements GoalVisitor<String>
 {
-	public static final GoalDescriptor INSTANCE = new GoalDescriptor();
+	public static final GoalInventoryDescriptor INSTANCE = new GoalInventoryDescriptor();
 	
 	@Override
 	public String visit(ItemGoal itemGoal)
